Trigger update with stateChange
